+++ packages/dumb/src/components/stats/insulin/insulin-stat.tsx
@@ -84,7 +84,8 @@ const InsulinStat: FunctionComponent<TotalInsulinStatProps> = (props) => {
   }
 
   const getPercentage = (value: number): string => {
-    const res = Math.round(100 * value / totalInsulin)
+    // We multiply by ten and divide by ten for the rounding
+    const res = Math.round(100 * 10 * value / totalInsulin) / 10
     return res > 0 ? res.toString(10) : EMPTY_DATA_PLACEHOLDER
   }
 
@@ -108,10 +109,11 @@ const InsulinStat: FunctionComponent<TotalInsulinStatProps> = (props) => {
           className={styles.boldValue}
         >
           <span className={styles.titleTotal}>
-            {totalInsulin}
-          </span>
-          <span className={styles.titleSuffix}>
-            U
+            <Chip
+              label={`${totalInsulin} ${t('insulin-unit-u')}`}
+              variant="outlined"
+              size="small"
+            />
           </span>
         </Box>
       </Box>
@@ -135,7 +137,7 @@ const InsulinStat: FunctionComponent<TotalInsulinStatProps> = (props) => {
                   label={`${entry.value > 0 ? entry.valueString : '0'} ${entry.units}`}
                   variant="outlined"
                   size="small"
-                  sx={{ marginRight: theme.spacing(2) }}
+                  sx={{ marginRight: theme.spacing(1) }}
                 />
                 <Box
                   className={styles.rowPercent}